pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ccenv"
version = "0.1.0"
description = "Claude Code environment manager - conda-like profile management for MCP, plugins, skills, and agents"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "hasibagen", email = "hasibagen@example.com"}
]
maintainers = [
    {name = "hasibagen", email = "hasibagen@example.com"}
]
keywords = ["claude", "claude-code", "mcp", "environment", "profile", "conda", "config"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/hasibagen/ccenv"
Documentation = "https://github.com/hasibagen/ccenv#readme"
Repository = "https://github.com/hasibagen/ccenv.git"
Issues = "https://github.com/hasibagen/ccenv/issues"
Changelog = "https://github.com/hasibagen/ccenv/blob/main/CHANGELOG.md"

[project.scripts]
ccenv = "ccenv.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ccenv = ["schemas/*.json"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"